Jitter: fix cast for 32 bits machines

+++ miasm2/jitter/arch/JitCore_x86.c
@@ -307,7 +307,7 @@ IMOD(64)
 
 
 
-void check_automod(JitCpu* jitcpu, uint64_t addr, int size)
+void check_automod(JitCpu* jitcpu, uint64_t addr, uint64_t size)
 {
 	PyObject *result;
